+++ src/glue/auth.cpp
@@ -1,0 +1,28 @@
+#include "glue/auth.hpp"
+
+namespace glue {
+
+auth::Permission PrivilegeToPermission(query::AuthQuery::Privilege privilege) {
+  switch (privilege) {
+    case query::AuthQuery::Privilege::MATCH:
+      return auth::Permission::MATCH;
+    case query::AuthQuery::Privilege::CREATE:
+      return auth::Permission::CREATE;
+    case query::AuthQuery::Privilege::MERGE:
+      return auth::Permission::MERGE;
+    case query::AuthQuery::Privilege::DELETE:
+      return auth::Permission::DELETE;
+    case query::AuthQuery::Privilege::SET:
+      return auth::Permission::SET;
+    case query::AuthQuery::Privilege::REMOVE:
+      return auth::Permission::REMOVE;
+    case query::AuthQuery::Privilege::INDEX:
+      return auth::Permission::INDEX;
+    case query::AuthQuery::Privilege::AUTH:
+      return auth::Permission::AUTH;
+    case query::AuthQuery::Privilege::STREAM:
+      return auth::Permission::STREAM;
+  }
+}
+
+}