+++ tests/manual/distributed_common.hpp
@@ -6,7 +6,7 @@
 #include "communication/result_stream_faker.hpp"
 #include "database/distributed_graph_db.hpp"
 #include "database/graph_db_accessor.hpp"
-#include "glue/conversion.hpp"
+#include "glue/communication.hpp"
 #include "query/interpreter.hpp"
 #include "query/typed_value.hpp"